requirements.txt
fastapi
uvicorn[standard]
redis
pillow
pytest
httpx
